+++ src/main/java/gift/model/Member.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.model;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
